+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.5.0 - released 13/02/2019
+ * Allow shared suggestion caching with `useGlobalCache` prop
  * More improvements to live docs
 
 # 0.4.0 - released 11/02/2019